Refatorar widget ItemAmigo
